feat: Add beverages overview and thumbnail
This commit introduces a new beverages overview screen and a thumbnail for the health overview.

Key changes:
- Added `BeverageOverviewView.kt` to display beverage data with navigation controls and a date range picker.
- Added `BeverageOverviewViewModel.kt` to manage beverage overview data and UI state.
- Added `PlotBeverages.kt` to display beverage data in a line or bar chart.
- Added `BeverageOverview.kt` data class for beverage overview items.
- Updated `HealthOverviewView.kt` to include the beverages thumbnail and navigate to `BeveragesOverview`.
- Renamed `StepsOverView` to `StepsOverview` in `HealthOverviewView.kt`.
- Updated `FluidUnit.kt` and `Beverage.kt` to use `UInt` for amount and adjust localized string generation.
- Updated `BeverageRepository.kt` to fetch and map beverage overview data, and changed `amount` to `UInt`.
- Updated `BeverageDao.kt` to rename `getTodayDrinks` to `getBeverages` and `upsertDrink` to `upsertBeverage`.
- Updated `BeverageViewModel.kt` to use `UInt` for amount and `targetBeverage`.
- Updated `RootScreen.kt` to handle navigation and title for `BeveragesOverview`.
- Updated `BottomBarNavGraph.kt` to include the `BeveragesOverview` screen.
- Updated `FitBeDatabase.kt` to use `upsertBeverage`.
- Updated `HealthOverviewViewModel.kt` to include `BeverageOverviewViewModel`.
- Updated `StepsViewModel.kt` to use `ProfileDefaults.STEPS` as a default target.
- Added new string resources for beverages overview and chart titles in English and German.
- Updated DI modules to include `BeverageOverviewViewModel`.

+++ composeApp/src/commonMain/kotlin/org/darthacheron/fitbe/health/beverages/Beverage.kt
@@ -30,17 +30,17 @@ data class Beverage(
     val id: Uuid,
     val profileId: Uuid,
     val dateUtc: Instant,
-    val amount: Int,
+    val amount: UInt,
     val beverage: String,
     val unit: FluidUnit
 ) {
     @Composable
     fun localizedString(): String {
         return when (unit) {
-            Cup -> pluralStringResource(Res.plurals.beverage_in_cup, quantity = amount, amount, beverage)
-            SmallGlass -> pluralStringResource(Res.plurals.beverage_in_small_glass, quantity = amount, amount, beverage)
-            NormalGlass -> pluralStringResource(Res.plurals.beverage_in_normal_glass, quantity = amount, amount, beverage)
-            LargeGlass -> pluralStringResource(Res.plurals.beverage_in_large_glass, quantity = amount, amount, beverage)
+            Cup -> pluralStringResource(Res.plurals.beverage_in_cup, quantity = amount.toInt(), amount, beverage)
+            SmallGlass -> pluralStringResource(Res.plurals.beverage_in_small_glass, quantity = amount.toInt(), amount, beverage)
+            NormalGlass -> pluralStringResource(Res.plurals.beverage_in_normal_glass, quantity = amount.toInt(), amount, beverage)
+            LargeGlass -> pluralStringResource(Res.plurals.beverage_in_large_glass, quantity = amount.toInt(), amount, beverage)
             Milliliter -> stringResource(Res.string.beverage_in_milliliter, amount, beverage)
             Centiliter -> stringResource(Res.string.beverage_in_centiliter, amount, beverage)
             Deciliter -> stringResource(Res.string.beverage_in_deciliter, amount, beverage)
